move tuner to ml
