+++ python/ray/ml/examples/pytorch/tune_torch_linear_dataset_example.py
@@ -3,8 +3,8 @@ import argparse
 import ray
 from ray import tune
 from ray.ml.train.integrations.torch import TorchTrainer
-from ray.tune.tune_config import TuneConfig
-from ray.tune.tuner import Tuner
+from ray.ml.tuner import Tuner
+from ray.ml.config import TuneConfig
 
 from torch_linear_dataset_example import train_func, get_datasets
